split web3j bean config
