+++ src/main/java/com/webank/webasebee/config/BeanConfig.java
@@ -15,33 +15,13 @@
  */
 package com.webank.webasebee.config;
 
-import java.io.File;
 import java.io.IOException;
-import java.io.InputStream;
-import java.util.ArrayList;
-import java.util.List;
-import java.util.concurrent.ConcurrentHashMap;
-import java.util.concurrent.ThreadPoolExecutor.AbortPolicy;
-import org.apache.commons.io.FileUtils;
-import org.apache.commons.lang3.StringUtils;
-import org.bcos.channel.client.Service;
-import org.bcos.channel.handler.ChannelConnections;
-import org.bcos.contract.tools.ToolConf;
-import org.bcos.web3j.crypto.Credentials;
-import org.bcos.web3j.crypto.GenCredential;
-import org.bcos.web3j.protocol.Web3j;
-import org.bcos.web3j.protocol.channel.ChannelEthereumService;
+
 import org.beetl.core.GroupTemplate;
 import org.beetl.core.resource.ClasspathResourceLoader;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-import org.springframework.scheduling.concurrent.ThreadPoolTaskExecutor;
 import org.springframework.transaction.annotation.EnableTransactionManagement;
-
-import com.google.common.collect.Lists;
-
-import lombok.extern.slf4j.Slf4j;
 
 /**
  * BeanConfig registers system common beans.
@@ -52,88 +32,7 @@ import lombok.extern.slf4j.Slf4j;
  */
 @Configuration
 @EnableTransactionManagement
-@Slf4j
 public class BeanConfig {
-    @Autowired
-    private SystemEnvironmentConfig systemEnvironmentConfig;
-
-    /**
-     * web3j thread pool config.
-     * 
-     * @return ThreadPoolTaskExecutor
-     */
-    @Bean
-    public ThreadPoolTaskExecutor getThreadPoolTaskExecutor() {
-        ThreadPoolTaskExecutor pool = new ThreadPoolTaskExecutor();
-        pool.setCorePoolSize(50);
-        pool.setMaxPoolSize(500);
-        pool.setQueueCapacity(500);
-        pool.setKeepAliveSeconds(30);
-        pool.setRejectedExecutionHandler(new AbortPolicy());
-        return pool;
-    }
-
-    /**
-     * web3j toolConf
-     * 
-     * @return ToolConf
-     */
-    @Bean
-    public ToolConf getToolConf() {
-        ToolConf toolConf = new ToolConf();
-        toolConf.setOutPutpath("./output/");
-        return toolConf;
-    }
-
-    /**
-     * web3j basic config.
-     * 
-     * @return Service
-     * @throws IOException
-     */
-    @Bean
-    public Service getService() throws IOException {
-        Service service = new Service();
-        service.setOrgID(systemEnvironmentConfig.getOrgId() + "");
-        ThreadPoolTaskExecutor threadPool = getThreadPoolTaskExecutor();
-        threadPool.setCorePoolSize(30);
-        service.setThreadPool(threadPool);
-        service.setConnectSeconds(10);
-        service.setConnectSleepPerMillis(10);
-        ChannelConnections con = new ChannelConnections();
-        ArrayList<String> list = new ArrayList<>();
-        String[] nodes = StringUtils.split(systemEnvironmentConfig.getNodeStr(), ";");
-        List<String> nodesList = Lists.newArrayList(nodes);
-        list.addAll(nodesList);
-        list.stream().forEach(s -> log.info("connect address: {}", s));
-        con.setConnectionsStr(list);
-
-        File caFile = loadFile(systemEnvironmentConfig.getConfigPath() + "ca.crt", "ca.crt");
-        con.setCaCertPath(caFile.getAbsolutePath());
-        log.info("Ca cert path is {}", caFile.getAbsolutePath());
-
-        File clientKeystoreFile =
-                loadFile(systemEnvironmentConfig.getConfigPath() + "client.keystore", "client.keystore");
-        con.setClientKeystorePath(clientKeystoreFile.getAbsolutePath());
-        log.info("clientKeystoreFile path is {}", clientKeystoreFile.getAbsolutePath());
-
-        con.setKeystorePassWord(systemEnvironmentConfig.getKeystorePassword());
-        con.setClientCertPassWord(systemEnvironmentConfig.getClientCertPassword());
-
-        ConcurrentHashMap<String, ChannelConnections> map = new ConcurrentHashMap<>();
-        map.put(systemEnvironmentConfig.getOrgId() + "", con);
-        service.setAllChannelConnections(map);
-        return service;
-    }
-
-    public File loadFile(String filePath, String fileName) throws IOException {
-        File file = new File(filePath);
-        if (!file.exists()) {
-            InputStream stream = getClass().getClassLoader().getResourceAsStream(fileName);
-            FileUtils.copyInputStreamToFile(stream, file);
-        }
-        return file;
-    }
 
     /**
      * Beetl render template.
@@ -147,38 +46,5 @@ public class BeanConfig {
         org.beetl.core.Configuration cfg = org.beetl.core.Configuration.defaultConfiguration();
         GroupTemplate gt = new GroupTemplate(resourceLoader, cfg);
         return gt;
-    }
-
-    /**
-     * Client Credentials, not used, but necessary.
-     * 
-     * @return Credentials
-     */
-    @Bean
-    public Credentials getCredentials() {
-        return GenCredential.create();
-    }
-
-    /**
-     * web3j run
-     * 
-     * @return Web3j
-     * @throws Exception
-     */
-    @Bean
-    public Web3j getWeb3j() throws Exception {
-        Service service = getService();
-        service.run();
-        Thread.sleep(3000);
-        log.info("开始启动web3j Service.");
-        log.info("初始化AOMP的ChannelEthereumService");
-        ChannelEthereumService channelEthereumService = new ChannelEthereumService();
-        // set all the timeout of sync request. Prevent the zombie thread.
-        channelEthereumService.setTimeout(30 * 1000);
-        log.info("初始化AOMP的ChannelEthereumService 结束");
-        channelEthereumService.setChannelService(service);
-        // 使用AMOP消息信道初始化web3j
-        Web3j web3j = Web3j.build(channelEthereumService);
-        return web3j;
     }
 }
